pyproject.toml
[tool.poetry]
name = "testasyncopenai"
version = "0.1.0"
description = ""
authors = ["thanglq150188 <thanglq150188@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.43.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
